+++ plugins/loginguard/smsapi/classes/Api/Response/Response.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SMSApi\Api\Response;
+
+interface Response
+{
+
+}
+