+++ plugins/loginguard/smsapi/classes/Exception/ActionException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SMSApi\Exception;
+
+class ActionException extends SmsapiException
+{
+
+}