+++ plugins/loginguard/smsapi/classes/Exception/InvalidParameterException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SMSApi\Exception;
+
+class InvalidParameterException extends SmsapiException
+{
+
+}