+++ plugins/loginguard/u2f/u2f.php
@@ -109,19 +109,21 @@ class PlgLoginguardU2f extends JPlugin
 
 		return array(
 			// Internal code of this TFA method
-			'name'          => $this->tfaMethodName,
+			'name'               => $this->tfaMethodName,
 			// User-facing name for this TFA method
-			'display'       => JText::_('PLG_LOGINGUARD_U2F_LBL_DISPLAYEDAS'),
+			'display'            => JText::_('PLG_LOGINGUARD_U2F_LBL_DISPLAYEDAS'),
 			// Short description of this TFA method displayed to the user
-			'shortinfo'     => JText::_('PLG_LOGINGUARD_U2F_LBL_SHORTINFO'),
+			'shortinfo'          => JText::_('PLG_LOGINGUARD_U2F_LBL_SHORTINFO'),
 			// URL to the logo image for this method
-			'image'         => 'media/plg_loginguard_u2f/images/u2f.svg',
+			'image'              => 'media/plg_loginguard_u2f/images/u2f.svg',
 			// Are we allowed to disable it?
-			'canDisable'    => true,
+			'canDisable'         => true,
 			// Are we allowed to have multiple instances of it per user?
-			'allowMultiple' => true,
+			'allowMultiple'      => true,
 			// URL for help content
-			'help_url' => $helpURL,
+			'help_url'           => $helpURL,
+			// Allow authentication against all entries of this TFA method. Otherwise authentication takes place against a SPECIFIC entry at a time.
+			'allowEntryBatching' => $this->params->get('allowEntryBatching', 1),
 		);
 	}
 
@@ -365,15 +367,44 @@ JS;
 		JText::script('PLG_LOGINGUARD_U2F_ERR_JS_INELIGIBLE_SIGN');
 		JText::script('PLG_LOGINGUARD_U2F_ERR_JS_TIMEOUT');
 
-		$options       = $this->_decodeRecordOptions($record);
-		$registrations = isset($options['registrations']) ? $options['registrations'] : array();
+		// Load the options from the record (if any), or from the entire method if the allowEntryBatching flag is set.
+		$registrations = $this->getRegistrations($record);
 
-		// If "Validate against all registered keys" is enabled we need to load all keys, not just the current one.
-		$u2fAuthData     = $this->u2f->getAuthenticateData($registrations);
-		$u2fAuthDataJSON = json_encode($u2fAuthData);
-
+		/**
+		 * The following code looks stupid. An explanation is in order.
+		 *
+		 * What we normally want to do is save the authentication data returned by getAuthenticateData into the session.
+		 * This is what is sent to the U2F key through the Javascript API and signed. The signature is posted back to
+		 * the form as the "code" which is read by onLoginGuardTfaValidate. That method will read the authentication
+		 * data from the session and pass it along with the key registration data (from the database) and the
+		 * authentication response (the "code" submitted in the form) to the U2F library for validation.
+		 *
+		 * Validation will work as long as the challenge recorded in the encrypted AUTHENTICATION RESPONSE matches, upon
+		 * decryption, the challenge recorded in the AUTHENTICATION DATA.
+		 *
+		 * I observed that for whatever stupid reason the browser was sometimes sending TWO requests to the server's
+		 * captive login page but only rendered the FIRST. This meant that the authentication data sent to the key had
+		 * already been overwritten in the session by the "invisible" second request. As a result the challenge would
+		 * not match and we'd get a validation error.
+		 *
+		 * The code below will attempt to read the authentication data from the session first. If it exists it will NOT
+		 * try to replace it (technically it replaces it with a copy of the same data - same difference!). If nothing
+		 * exists in the session, however, it WILL store the (random seeded) result of the getAuthenticateData method.
+		 * Therefore the first request to the captive login page will store a new set of authentication data whereas the
+		 * second, "invisible", request will just reuse the same data as the first request, fixing the observed issue in
+		 * a way that doesn't compromise security.
+		 *
+		 * In case you are wondering, yes, the data is removed from the session in the onLoginGuardTfaValidate method.
+		 * In fact it's the first thing we do after reading it, preventing constant reuse of the same set of challenges.
+		 *
+		 * That was fun to debug - for "poke your eyes with a rusty fork" values of fun.
+		 */
 		$session = JFactory::getSession();
-		$session->set('u2f.authentication', serialize($u2fAuthData), 'com_loginguard');
+		$u2fAuthData     = $this->u2f->getAuthenticateData($registrations);
+		$u2fAuthData     = $session->get('u2f.authentication', base64_encode(serialize($u2fAuthData)), 'com_loginguard');
+		$u2fAuthData     = unserialize(base64_decode($u2fAuthData));
+		$u2fAuthDataJSON = json_encode($u2fAuthData);
+		$session->set('u2f.authentication', base64_encode(serialize($u2fAuthData)), 'com_loginguard');
 
 		$js = <<< JS
 window.jQuery(document).ready(function($) {
@@ -397,21 +428,23 @@ JS;
 
 		return array(
 			// Custom HTML to display above the TFA form
-			'pre_message'  => JText::_('PLG_LOGINGUARD_U2F_LBL_INSTRUCTIONS'),
+			'pre_message'        => JText::_('PLG_LOGINGUARD_U2F_LBL_INSTRUCTIONS'),
 			// How to render the TFA code field. "input" (HTML input element) or "custom" (custom HTML)
-			'field_type'   => 'custom',
+			'field_type'         => 'custom',
 			// The type attribute for the HTML input box. Typically "text" or "password". Use any HTML5 input type.
-			'input_type'   => '',
+			'input_type'         => '',
 			// Placeholder text for the HTML input box. Leave empty if you don't need it.
-			'placeholder'  => '',
+			'placeholder'        => '',
 			// Label to show above the HTML input box. Leave empty if you don't need it.
-			'label'        => '',
+			'label'              => '',
 			// Custom HTML. Only used when field_type = custom.
-			'html'         => $html,
+			'html'               => $html,
 			// Custom HTML to display below the TFA form
-			'post_message' => '',
+			'post_message'       => '',
 			// URL for help content
-			'help_url'     => $helpURL,
+			'help_url'           => $helpURL,
+			// Allow authentication against all entries of this TFA method. Otherwise authentication takes place against a SPECIFIC entry at a time.
+			'allowEntryBatching' => $this->params->get('allowEntryBatching', 1),
 		);
 	}
 
@@ -445,9 +478,8 @@ JS;
 			return false;
 		}
 
-		// Load the options from the record (if any)
-		$options = $this->_decodeRecordOptions($record);
-		$registrations = isset($options['registrations']) ? $options['registrations'] : array();
+		// Load the options from the record (if any), or from the entire method if the allowEntryBatching flag is set.
+		$registrations = $this->getRegistrations($record);
 
 		// Get the authentication response
 		$authenticateResponse = json_decode($code);
@@ -468,7 +500,7 @@ JS;
 			return false;
 		}
 
-		$authenticationRequest = unserialize($authenticationRequest);
+		$authenticationRequest = unserialize(base64_decode($authenticationRequest));
 
 		if (empty($authenticationRequest))
 		{
@@ -658,5 +690,51 @@ JS;
 		}
 
 		return $return;
+	}
+
+	/**
+	 * Get the security key registrations for a given record. If the allowEntryBatching flag is 0 (No) we only return
+	 * the key registrations for the given record. If the allowEntryBatching flag is 1 (Yes) we return the combined key
+	 * registrations for all security key records of the user ID found in the $record object.
+	 *
+	 * @param   stdClass  $record  The LoginGuard record
+	 *
+	 * @return  array  Security key registrations for use by the U2F library
+	 */
+	private function getRegistrations($record)
+	{
+		$options       = $this->_decodeRecordOptions($record);
+
+		if ($this->params->get('allowEntryBatching', 1) == 0)
+		{
+			return isset($options['registrations']) ? $options['registrations'] : array();
+		}
+
+		$registrations = array();
+
+		try
+		{
+			$db = JFactory::getDbo();
+			$query = $db->getQuery(true)
+			            ->select('*')
+			            ->from($db->qn('#__loginguard_tfa'))
+			            ->where($db->qn('user_id') . ' = ' . $db->q($record->user_id))
+			            ->where($db->qn('method') . ' = ' . $db->q($record->method));
+			$records = $db->setQuery($query)->loadObjectList();
+		}
+		catch (Exception $e)
+		{
+			$records = array();
+		}
+
+		// Loop all records, stop if at least one matches
+		foreach ($records as $aRecord)
+		{
+			$recordOptions       = $this->_decodeRecordOptions($aRecord);
+			$recordRegistrations = isset($recordOptions['registrations']) ? $recordOptions['registrations'] : array();
+			$registrations       = array_merge($registrations, $recordRegistrations);
+		}
+
+		return $registrations;
 	}
 }